fix(ui): preserve fractional responsive bounds

+++ composeApp/src/commonMain/kotlin/ru/sodovaya/volty/presentation/dashboard/DashboardScreen.kt
@@ -101,7 +101,7 @@ fun DashboardScreen(component: DashboardComponent) {
             .fillMaxSize()
             .statusBarsPadding()
     ) {
-        val layoutMode = responsiveLayoutMode(maxWidth.value.toInt(), maxHeight.value.toInt())
+        val layoutMode = responsiveLayoutMode(maxWidth.value, maxHeight.value)
         Column(
             modifier = Modifier
                 .fillMaxSize()
