+++ composeApp/src/commonMain/kotlin/ru/sodovaya/volty/presentation/ride/RideDashboardScreen.kt
@@ -161,7 +161,7 @@ fun RideDashboardScreen(component: RideDashboardComponent) {
             .fillMaxSize()
             .statusBarsPadding()
     ) {
-        val layoutMode = responsiveLayoutMode(maxWidth.value.toInt(), maxHeight.value.toInt())
+        val layoutMode = responsiveLayoutMode(maxWidth.value, maxHeight.value)
         Column(
             modifier = Modifier
                 .fillMaxSize()
